feat: seperate circular loading state indicator and global loading indicator
