+++ apps/discord-app-vite/src/components/GlobalLoading/assets/loading_icon.mp4
